fix: resolve ambiguous Roaring64Map overload on GCC/Linux
Cast loop variable to uint64_t when calling Roaring64Map::contains()
and add() in expandContext().  On GCC, 'unsigned long long' (from 0ULL)
is distinct from both uint32_t and uint64_t overloads, causing an
ambiguous call error with -Werror.

+++ src/logdata/src/logfiltereddata.cpp
@@ -263,8 +263,8 @@ void LogFilteredData::rebuildContextLines()
                                    : 0ULL;
             const auto end = std::min( line + static_cast<uint64_t>( p->n ), p->maxLine - 1 );
             for ( auto i = start; i <= end; ++i ) {
-                if ( !p->baseSet->contains( i ) ) {
-                    p->result->add( i );
+                if ( !p->baseSet->contains( static_cast<uint64_t>( i ) ) ) {
+                    p->result->add( static_cast<uint64_t>( i ) );
                 }
             }
             return true;
